+++ src/anonymizer/view/index.py
@@ -14,7 +14,7 @@ from anonymizer.model.anonymizer import AnonymizerModel, PHI_IndexRecord
 from anonymizer.utils.translate import _
 from anonymizer.view.dashboard import Dashboard
 from anonymizer.view.delete_studies_dialog import DeleteStudiesDialog
-from anonymizer.view.pixels import PixelsView
+from anonymizer.view.projection import ProjectionView
 
 logger = logging.getLogger(__name__)
 
@@ -40,15 +40,15 @@ class IndexView(tk.Toplevel):
         self,
         parent: Dashboard,
         project_controller: ProjectController,
-        mono_font: ctk.CTkFont,
+        char_width_px: int,
     ):
         super().__init__(master=parent)
-        self._char_width_px = mono_font.measure("A")
+        self._char_width_px = char_width_px
         self._parent = parent
         self._controller = project_controller
         self._anon_model: AnonymizerModel = project_controller.anonymizer.model
         self._phi_index: list[PHI_IndexRecord] | None = None
-        self._pixels_view: PixelsView | None = None
+        self._pixels_view: ProjectionView | None = None
 
         self.title("View PHI Index")
         self.resizable(True, True)
@@ -112,8 +112,8 @@ class IndexView(tk.Toplevel):
         self._view_pixels_button = ctk.CTkButton(
             self._button_frame,
             width=ButtonWidth,
-            text=_("View Pixels"),
-            command=self._view_pixels_button_pressed,
+            text=_("View Projections"),
+            command=self._view_projections_button_pressed,
         )
         self._view_pixels_button.grid(row=0, column=4, padx=PAD, pady=PAD, sticky="w")
 
@@ -190,7 +190,7 @@ class IndexView(tk.Toplevel):
         for row, record in enumerate(self._phi_index):
             self._tree.insert("", 0, iid=row, values=record.flatten())
 
-    def _view_pixels_button_pressed(self):
+    def _view_projections_button_pressed(self):
         if self._phi_index is None:
             return
 
@@ -202,10 +202,14 @@ class IndexView(tk.Toplevel):
 
         rows_selected = list(self._tree.selection())
         logger.info(f"View Pixels button pressed, {len(rows_selected)} studies selected")
+        if len(rows_selected) == 0:
+            logger.error("No Studies selected")
+            return
+
         selected_indices = [int(row) for row in rows_selected]  # convert to integers
         selected_phi_records = [self._phi_index[i] for i in selected_indices]
 
-        self._pixels_view = PixelsView(self, self._controller.model.images_dir(), selected_phi_records)
+        self._pixels_view = ProjectionView(self, self._controller.model.images_dir(), selected_phi_records)
         if self._pixels_view is None:
             logger.error("Internal Error creating PixelsView")
             return